refactor : 양방향 친구 추가가 가능하도록 로직 변경

+++ src/main/java/com/example/newsbara/user/controller/FollowController.java
@@ -47,7 +47,7 @@ public class FollowController {
     @Operation(summary = "친구 수락/거절 API",
             description = "자신에게 온 친구 요청을 거절/수락하는 API입니다.")
     public ApiResponse<HandleResDto> handleRequest(
-            @PathVariable Long requestId,
+            @PathVariable Integer requestId,
             @RequestBody HandleReqDto request,
             Principal principal) {
